Adjsuted mass rebuild to work with pre-changed packages

+++ scritps/massRebuild/massBuild.sh
@@ -15,10 +15,10 @@ readonly SCRIPT_DIR="$( cd -P "$( dirname "$SCRIPT_SOURCE" )" && pwd )"
 
 set -e
 set -o pipefail
-branch=f40
-TAG=f40-build-side-84857
-targettedSelection="CFR"
-DO="false"
+branch=rawhide
+TAG=rawhide
+##targettedSelection="CFR"
+
 
 FILE_WITH_PKGS="$SCRIPT_DIR/../fillCopr/exemplarResults/depndent-packages.jbump"
 
@@ -30,10 +30,10 @@ echo " * ${!pkgs[@]}"
 echo "type yes and enter"
 read
 if [ ! "x${REPLY}" = "xyes" ] ; then exit 1 ; fi
-#lets add few more which builds to long. You will handle them manually alter
-for k in icecat rstudio ceph hdf5 libreoffice chromium ; do
-  pkgs[$k]=unused_now
-done
+#lets add few more which builds to long. You will handle them manually later
+#for k in icecat rstudio ceph hdf5 libreoffice chromium ; do
+#  pkgs[$k]=unused_now
+#done
 regex=`echo "${!pkgs[@]}" | sed  's;\s\+;$ -e ^;g'`
 regex=" -e ^someNonsnense$regex$"
 echo "total known packages: "
@@ -44,7 +44,7 @@ echo -n " * "
 cat $FILE_WITH_PKGS | grep  -v $regex | grep "$targettedSelection" | wc -l
 echo "dont forget to handle following packages manually!"
 echo " * " ${!pkgs[@]}
-RESULTS_DIR="$PWD/results"
+RESULTS_DIR="$SCRIPT_DIR/results"
 echo "DO=$DO targettedSelection=$targettedSelection TAG=$TAG"
 echo "Are you kinit into FEDORAPROJECT.ORG? Are you proven packager? Do you really wish to run in branch $branch?"
 echo "type yes and enter. If you are not proven packager, it will fail on foreign pkgs. This will discard all old results $RESULTS_DIR"
@@ -56,20 +56,29 @@ echo "starting mass rebuild for $TAG"
 set +e # disputable... but yo do not want to lost 24h long script becasue of minute network issue i second hour
 #inspired by https://github.com/hroncok/mini-mass-rebuild ... fedpkg-bump-build.sh I think
 # if you wish to run it on some set of pkgs - eg on the keys from above - copy this script, and replace below `cat...` by hardcoded list
+
 for pkg in `cat $FILE_WITH_PKGS | grep  -v $regex  | grep "$targettedSelection"` ; do 
-  fedpkg clone $pkg  2>&1 | tee $RESULTS_DIR/${pkg}.log
+  if [ "x$CLONE" == "xfalse" ] ; then
+    ls -l -d $pkg
+  else
+    fedpkg clone $pkg  2>&1 | tee $RESULTS_DIR/${pkg}.log
+  fi
   pushd $pkg
     git checkout $branch
-    MSG_TITLE="Rebuilt for java-21-openjdk as system jdk"
+    MSG_TITLE="Rebuilt for java-25-openjdk as preffered jdk"
     MSG="$MSG_TITLE
 
-https://fedoraproject.org/wiki/Changes/Java21
+https://fedoraproject.org/wiki/Changes/Java25AndNoMoreSystemJdk
 "
     rpmdev-bumpspec -c "$MSG_TITLE" $pkg.spec | tee -a $RESULTS_DIR/${pkg}.log
     git commit --allow-empty ${pkg}.spec -m "$MSG" | tee -a $RESULTS_DIR/${pkg}.log
-    if [ "x$DO" == "xtrue" ] ; then
-      git push | tee -a $RESULTS_DIR/${pkg}.log
-      fedpkg build  --target $TAG --fail-fast --nowait --background 2>&1 | tee -a $RESULTS_DIR/${pkg}.log
+    if [ "x$DO" == "xtrue" -o "x$DO" == "xscratch" ] ; then
+      if [ "x$DO" == "xscratch" ] ; then
+        fedpkg build  --target $TAG --fail-fast --nowait --background --scratch --srpm 2>&1 | tee -a $RESULTS_DIR/${pkg}.log
+      else
+        git push | tee -a $RESULTS_DIR/${pkg}.log
+        fedpkg build  --target $TAG --fail-fast --nowait --background 2>&1 | tee -a $RESULTS_DIR/${pkg}.log
+      fi
       TASK=`cat $RESULTS_DIR/${pkg}.log  | grep "Created task: " | sed "s/.*: //"`
       koji watch-task $TASK >> $RESULTS_DIR/${pkg}.log &
       sleep 20
@@ -78,7 +87,11 @@ https://fedoraproject.org/wiki/Changes/Java21
       rpmbuild --rebuild $pkg*.src.rpm 2>&1 | tee -a $RESULTS_DIR/${pkg}.log # this is nearly irrelevant, as it do not even use sidetag but good enough to verify commit and so
     fi
   popd
-  rm -rf $pkg
+  if [ "x$CLONE" == "xfalse" ] ; then
+    ls -l -d $pkg
+  else
+    rm -rf $pkg
+  fi
   processes=`ps | wc -l`
   while [ $processes -gt 60 ] ; do 
     processes=`ps | wc -l`
